Electrode_Wise_Comparison/ModelFinetuning/VisionTransformerModel/requirements.txt
addict==2.4.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
attrs==22.1.0
autopep8==1.7.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
einops==0.4.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.0
fonttools==4.37.1
huggingface-hub==0.9.1
idna==3.3
imageio==2.21.2
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.17.0
jupyter-server==1.23.2
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
monai==0.9.1
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.6
nibabel==4.0.2
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.2
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parameterized==0.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.4
prometheus-client==0.15.0
prompt-toolkit==3.0.32
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
pydicom==2.3.0
Pygments==2.13.0
pylibjpeg==1.4.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-gdcm==3.0.17.1
pytz==2022.2.1
PyWavelets==1.4.0
PyYAML==6.0
pyzmq==24.0.1
regex==2022.9.13
requests==2.28.1
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
Send2Trash==1.8.0
SimpleITK==2.2.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.1
terminado==0.17.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2022.8.12
timm==0.4.12
tinycss2==1.2.1
tokenizers==0.12.1
toml==0.10.2
# torch==1.12.1+cu116
# torchvision==0.13.1+cu116
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.21.3
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
yapf==0.32.0
